chore(mv): moved util test files into separate folder
