+++ test/utils/get-filesystem.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import * as fs from 'node:fs';
-import { getFilesystem } from '../src/utils/get-filesystem';
+import { getFilesystem } from '../../src/utils/get-filesystem';
 
 describe('getFilesystem', () => {
     it('should return inmemory filesystem during test', () => {
